requirements/base.txt
asgiref==3.11.1
Django==5.0
django-ratelimit==4.1.0
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
pillow==12.1.1
PyJWT==2.11.0
python-decouple==3.8
redis==7.1.1
sqlparse==0.5.5
python-decouple
Pillow
djangorestframework
djangorestframework-simplejwt

redis
drf-spectacular
httpx
adrf
django-parler
pytz
channels 
channels-redis

celery
flower

